+++ config/deploy.yml
@@ -41,6 +41,11 @@ env:
     - ADMIN_USERNAME
     - ADMIN_PASSWORD
     - CINEGRAPH_API_KEY
+    - CLOUDFLARE_ACCOUNT_ID
+    - CLOUDFLARE_ACCESS_KEY_ID
+    - CLOUDFLARE_SECRET_ACCESS_KEY
+    - R2_BUCKET
+    - R2_CDN_URL
 
 ssh:
   user: holden
